web/requirements.txt
Flask
flask_restful
pymongo
PyJWT
bcrypt
numpy
tensorflow
requests
